### 2025-05-04 #8005 @csoneson
👋🏻 @pr4deepr, @Armavica, @jrkerns - would you be interested in reviewing this submission for JOSS?

> scmorph: a Python package for analysing single-cell morphological profiles
#8005

Thanks in advance!
