### 2025-05-05 #8005 @Armavica
> 👋🏻 [@pr4deepr](https://github.com/pr4deepr), [@Armavica](https://github.com/Armavica), [@jrkerns](https://github.com/jrkerns) - would you be interested in reviewing this submission for JOSS?
>
> > scmorph: a Python package for analysing single-cell morphological profiles
> > [#8005](https://github.com/openjournals/joss-reviews/issues/8005)
>
> Thanks in advance!

Hi, I would be happy to review this submission but I am a bit overwhelmed at the moment. I can try but I am not sure how much time I will have this month. I will definitely have more time in June, though.
