### 2025-05-15 #8005 @csoneson
👋🏻 @Emptymalei, @pr4deepr, @jrkerns, @bgailleton - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> scmorph: a Python package for analysing single-cell morphological profiles
#8005

You can find more information about our checklist-driven review process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Thanks in advance!
