### 2025-05-15 #8005 @jrkerns
My apologies for the late reply. I don't currently have the bandwidth.
Thank you for the consideration!

On Thu, May 15, 2025 at 3:03 PM Charlotte Soneson ***@***.***>
wrote:

> *csoneson* left a comment (openjournals/joss-reviews#8005)
> <https://github.com/openjournals/joss-reviews/issues/8005#issuecomment-2884929415>
>
> 👋🏻 @emptymalei <https://github.com/emptymalei>, @pr4deepr
> <https://github.com/pr4deepr>, @jrkerns <https://github.com/jrkerns>,
> @bgailleton <https://github.com/bgailleton> - would you be interested in
> reviewing this submission for the Journal of Open Source Software (JOSS)?
>
> scmorph: a Python package for analysing single-cell morphological profiles
> #8005 <https://github.com/openjournals/joss-reviews/issues/8005>
>
> You can find more information about our checklist-driven review process
> here <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
>
> Thanks in advance!
>
