### 2025-05-16 #8005 @csoneson
Thanks @bgailleton - that sounds great.

@tushardave26 - do you feel comfortable commenting on the scientific aspect of the submission (to the extent covered by the JOSS [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html))? I am also happy to look for a third reviewer.
