### 2025-05-19 #8005 @bgailleton
> Thanks @bgailleton - that sounds great.
>
> @tushardave26 - do you feel comfortable commenting on the scientific aspect of the submission (to the extent covered by the JOSS [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html))? I am also happy to look for a third reviewer.
>

@csoneson Hi, I'm so sorry, but my baby is arriving a bit earlier than expected, and I won't be able to commit to this review for at least a month.
