### 2025-05-19 #8005 @tushardave26
> [@tushardave26](https://github.com/tushardave26) - do you feel comfortable commenting on the scientific aspect of the submission (to the extent covered by the JOSS [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html))? I am also happy to look for a third reviewer.

@csoneson - Yes, I can take a look at that.
