### 2025-06-10 edbiomedai/scmorph#7 @Armavica
Clarify documentation for Batch Effects
### Description of feature

Hi, as part of the review of this JOSS submission (https://github.com/openjournals/joss-reviews/issues/8324), I am trying to follow the tutorials on your documentation.
On the [Batch Effects page](https://scmorph.readthedocs.io/en/latest/tutorials/basics/batch_effects.html), I am wondering where the variable names come from:
- I saw that `"Image_Metadata_Plate"` is a metadata associated to `adata.obs`
- However, I couldn't find `"Cells_Intensity_MeanIntensity_AGP"` neither in `adata.obs` nor in `adata.var`: where does this one come from?
- I also don't understand how you plotted the PCA graph. Is this another variable of the dataset or is it computed manually?
