### 2025-06-11 edbiomedai/scmorph#8 @tushardave26
Specific Suggestions for Scientific Enhancement
### Description of feature

Hi @jeskowagner, as part of my review of the JOSS submission (https://github.com/openjournals/joss-reviews/issues/8324), I’ve outlined several suggestions for scientific enhancement. These are organized into two categories: (1) **Mandatory** and (2) **Optional** recommendations.

### Mandatory

1. **Quantitative Benchmarking:** While the manuscript effectively outlines `scMorph`'s distinct features and advantages, quantitative benchmarking against existing tools (e.g., `pycytominer`, `SPACe`) would significantly bolster the scientific claims.
    * **Suggestion:** Consider including a section or a supplementary figure demonstrating `scMorph`'s performance (e.g., in terms of runtime, memory usage, or specific metrics like batch correction efficacy or hit calling accuracy) on shared or publicly available datasets. For instance, demonstrate how `scMorph`'s `scone`-based batch correction quantitatively reduces batch effects while preserving biological signal, compared to other methods, or compare hit calling accuracy on a dataset with known ground truth. This would provide concrete scientific evidence for the improvements in scalability and method performance described.
2. **Elobaration on "Interpretability of Features" in Batch Correction:** The manuscript states that `scone` "retains interpretability of features". This is a crucial scientific advantage for morphological profiling data, where feature interpretation is often key to biological insights.
    * **Suggestion:** Expand on this point in the manuscript or documentation. A brief scientific explanation of why `scone` maintains interpretability better than other common batch correction methods (e.g., how it avoids aggressive transformations that obscure feature meaning) would be highly valuable. Perhaps illustrate this with a small example in the documentation showing how original feature relationships are better preserved after `scone` correction.
3. **Guidance on Feature Selection Strategies:** `scMorph` offers integrated feature selection methods, including those based on `Chatterjee correlation coefficient` or `Kruskal-Wallis test`.
    * **Suggestion:** Provide scientific guidance in the documentation on when to use which feature selection method. For instance, are certain methods more appropriate for specific data distributions, biological questions, or types of confounding factors? This would help users make informed scientific choices during preprocessing.

### Optional

1. **Deeper Dive into "Single-cell Specific Methods":** The "Statement of Need" mentions the lack of open-source software offering "single-cell specific methods". While the manuscript and documentation mention not requiring cell averaging and single-cell outlier detection, a more explicit scientific discussion would be beneficial.
    * **Suggestion:** Provide a more precise definition or conceptual explanation of what constitutes a "single-cell specific method" in the context of morphological profiling. Briefly contrast this with methods that average features across cells, and explain why retaining single-cell resolution is scientifically superior for uncovering subtle cellular heterogeneity and rare phenotypes that might be missed by population-level approaches.
3. **Examples for Multi-nucleated Cell Analysis:** The manuscript notes `scMorph`'s capability to analyze multi-nucleated cells. This is a niche but essential feature for specific biological contexts.
    * **Suggestion:** If possible, include a small example or a dedicated section in the documentation/tutorials highlighting the use case for multi-nucleated cell analysis. Describe a scientific scenario where this feature is particularly valuable (e.g., differentiating cells, cell fusion, or specific disease models) to guide users on its application.
3. **Discussion of Limitations and Assumptions:** For robust scientific software, acknowledging potential limitations or assumptions of the integrated methods is good practice.
    * **Suggestion:** Consider adding a brief section in the documentation (e.g., "Considerations" or "Limitations") that discusses any known limitations of `scMorph` or the underlying statistical methods. For example, what are the key assumptions of the `Kolmogorov-Smirnov` statistic for hit calling or `slingshot` for trajectory inference? This would help users interpret results more accurately and understand when `scMorph` might be less suitable.
