### 2025-06-11 edbiomedai/scmorph#9 @tushardave26
Potential reference duplication
### Description of feature

Hi @jeskowagner, as part of my review of the JOSS submission (https://github.com/openjournals/joss-reviews/issues/8324), I have read the manuscript carefully and found a potential reference duplication.

You cite both:
- Bézieux et al. (2021) in bioRxiv
- Roux de Bézieux et al. (2024) in Nature Communications

These appear to refer to the same work at different stages of publication. In such cases, it is generally recommended to cite only the peer-reviewed, published version of the work. I would suggest that you clarify whether both references are necessary or update the citation to reference only the final published article.
