### 2025-06-11 edbiomedai/scmorph#7 @tushardave26
@jeskowagner - I agree with the point raised by @Armavica. I also could not find the observation `Cells_Intensity_MeanIntensity_AGP` in either the `adata` or `adata_ctrl` `AnnData` objects. Could you please clarify where this observation is being accessed from?

Additionally, it seems that the tutorial is missing the command that generates the `PCA` plot referenced in the narrative. Including that line would help ensure reproducibility and clarity.
