### 2025-06-17 e16741b
fix ref duplication (#9)
