+++ paper/paper_minimal.bib
@@ -1,21 +1,4 @@
 
-@misc{BezieuxEtAl2021,
-  title         = {Trajectory Inference across Multiple Conditions with Condiments: Differential Topology, Progression, Differentiation, and Expression},
-  shorttitle    = {Trajectory Inference across Multiple Conditions with Condiments},
-  author        = {de B{\'e}zieux, Hector Roux and den Berge, Koen Van and Street, Kelly and Dudoit, Sandrine},
-  year          = {2021},
-  month         = mar,
-  primaryclass  = {New Results},
-  pages         = {2021.03.09.433671},
-  publisher     = {bioRxiv},
-  doi           = {10.1101/2021.03.09.433671},
-  urldate       = {2022-12-27},
-  archiveprefix = {bioRxiv},
-  chapter       = {New Results},
-  copyright     = {{\copyright} 2021, Posted by Cold Spring Harbor Laboratory. This pre-print is available under a Creative Commons License (Attribution 4.0 International), CC BY 4.0, as described at http://creativecommons.org/licenses/by/4.0/},
-  langid        = {english}
-}
-
 @article{ColeEtAl2019,
   title   = {Performance {{Assessment}} and {{Selection}} of {{Normalization Procedures}} for {{Single-Cell RNA-Seq}}},
   author  = {Cole, Michael B. and Risso, Davide and Wagner, Allon and DeTomaso, David and Ngai, John and Purdom, Elizabeth and Dudoit, Sandrine and Yosef, Nir},
