### 2025-06-20 edbiomedai/scmorph#8 @jeskowagner
Further, on quantitative benchmarking, there now is a [notebook](https://scmorph.readthedocs.io/en/latest/notebooks/scalability.html) that demonstrates that scmorph scales better to many cells than pytcytominer.
