### 2025-06-20 4de1477
exclude non-tutorial notebooks from linting
