### 2025-06-20 8fad491
add brief explainer on interpretability
