### 2025-06-20 e8c73ab
clarify feature selection in docstrings
