### 2025-06-20 edbiomedai/scmorph#8 @jeskowagner
On the third mandatory point, I have added a bit more description to the docstrings [here](https://scmorph.readthedocs.io/en/latest/api/generated/scmorph.pp.kruskal_test.html#scmorph.pp.kruskal_test) and [here](https://scmorph.readthedocs.io/en/latest/api/generated/scmorph.pp.select_features.html#scmorph.pp.select_features). In short, Kruskwal-Wallis is used to flag features associated with plate/batch effects, whereas `select_features` flags correlated features which may be undesirable (and is done more to stay consistent with the field).
