### 2025-06-24 edbiomedai/scmorph#8 @jeskowagner
To summarise:

**Mandatory**
- [x] _Quantitative Benchmarking_. See [new tutorial](https://scmorph.readthedocs.io/en/latest/tutorials/basics/why_scone.html) and [showcase](https://scmorph.readthedocs.io/en/latest/notebooks/scalability.html) of scalability.
- [x] _Elobaration on "Interpretability of Features" in Batch Correction_. [Paper](https://github.com/edbiomedai/scmorph/blob/paper/paper/paper.pdf) has been amended accordingly.
- [x] _Guidance on Feature Selection Strategies_. Docs have been updated as per my [comment](https://github.com/edbiomedai/scmorph/issues/8#issuecomment-2991334557).

**Optional**
- [x] _Deeper Dive into "Single-cell Specific Methods"_. This has been made a bit more clear in the paper now.
- [x]  _Examples for Multi-nucleated Cell Analysis_. Unfortunately, no published data containing multi-nucleated profiles exist. While we have used `scmorph` to this end internally, it will be difficult to substantiate this claim until the accompanying data have been released. I have therefore amended the manuscript to remove the mention of multi-nucleated cells.
- [x] _Discussion of Limitations and Assumptions_. I have added a brief discussion of this in the manuscript.

@tushardave26, could you please look over these changes and let me know whether they satisfy your concerns?

Thanks! Best,

Jesko
