### 2025-06-25 edbiomedai/scmorph#8 @tushardave26
> I have also updated the paper to briefly mention that the `scone`-based strategy does not require z-score transformation. To expand on this, if looking at cell sizes for example, after `scone` correction statements such as "treatment X reduced cell diameter by 30% (5 μM)" remain possible. Whereas the other batch correction techniques rely on z-score transformations, which prohibit such interpretations.

@jeskowagner – I don’t see any mention of the changes described above reflected in the manuscript, and it appears the manuscript has not yet been updated. If I’m overlooking something, please feel free to point me to the exact location where the changes have been made.
