### 2025-06-25 edbiomedai/scmorph#8 @tushardave26
> On the third mandatory point, I have added a bit more description to the docstrings [here](https://scmorph.readthedocs.io/en/latest/api/generated/scmorph.pp.kruskal_test.html#scmorph.pp.kruskal_test) and [here](https://scmorph.readthedocs.io/en/latest/api/generated/scmorph.pp.select_features.html#scmorph.pp.select_features). In short, Kruskwal-Wallis is used to flag features associated with plate/batch effects, whereas `select_features` flags correlated features which may be undesirable (and is done more to stay consistent with the field).

@jeskowagner - The `docstring` for the `Kruskal-Wallis` method seems to be missing some context.

> Note that while it does reduce feature space, its main purpose is to untrustworthy features associated with technical confounders.

It looks like there may be a missing word before “to untrustworthy.” Please let me know if I’m misunderstanding anything.
