### 2025-06-25 edbiomedai/scmorph#10 @tushardave26
ImplicitModificationWarning Raised During Execution of Batch Effects Tutorial
### Description of feature

Hi @jeskowagner, as part of my review of the JOSS submission (https://github.com/openjournals/joss-reviews/issues/8324), I've been trying to run all the steps described in the Batch Effect tutorial. While doing so, I encountered `ImplicitModificationWarning` specifically in two code block execution as list below:

**Code Block - 1**
```
adata_corrected = sm.pp.remove_batch_effects(
    adata,
    batch_key="Image_Metadata_Plate",
    treatment_key="TARGETGENE",
    control="UNTREATED",
    copy=True,
)
```
**Warning from Code Block - 1**
```
scmorph/utils/utils.py#line=149): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata[idx].X = fun(X, group) if takes_group else fun(X)
```
**Code Block - 2**
```
adata_corrected = adata.copy()

# Perform per-plate z-scoring
sm.pp.scale_by_batch(
    adata_corrected,
    batch_key="Image_Metadata_Plate",
    treatment_key="TARGETGENE",
    control="UNTREATED",
)

adata_ctrl = adata_corrected[adata_corrected.obs["TARGETGENE"] == "UNTREATED"]
sm.pl.cumulative_density(
    adata_ctrl,
    "Cells_Intensity_MeanIntensity_AGP",
    color="Image_Metadata_Plate",
    xlim=(-1, 1),
    palette="viridis",
    xlabel="Avg. AGP intensity [z-score]",
)
plt.show()
```
**Warnings from Code Block - 2**
```
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
```
`ImplicitModificationWarning` is a **warning class** typically raised by Python packages like `anndata` or `scanpy`, when you attempt to modify an `AnnData` **object in-place** in a way that implicitly changes the structure or contents of its internal components (e.g., .obs, .var, .X, .layers, etc.), without explicitly copying or declaring the change.

I encourage you to try running your `Batch Effects` tutorial as described in the documentation to see if you can reproduce the warnings. If the warnings do appear, it may be worth updating the tutorial to help users avoid them.
