### 2025-06-25 edbiomedai/scmorph#10 @tushardave26
@jeskowagner - I encountered the same warning message for the code block described below as well.

```
sm.pp.pca(adata_ctrl)

sm.pl.cumulative_density(
    adata_ctrl,
    0,
    "pca",
    color="Image_Metadata_Plate",
    xlim=(-20, 20),
    palette="viridis",
    xlabel="PC1 coordinate",
)
```
**Warning message**
```
scanpy/preprocessing/_pca/__init__.py:383](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scanpy/preprocessing/_pca/__init__.py#line=382): ImplicitModificationWarning: Setting element `.obsm['X_pca']` of view, initializing view as actual.
  adata.obsm[key_obsm] = X_pca
```
