### 2025-06-26 edbiomedai/scmorph#10 @tushardave26
@jeskowagner - In the `Hit Calling` tutorial, I encountered the same warning for the following code blocks.

**Code Block - 1**
```
# Step 1:
adata = load_example_data()

# Compute PCA
sm.pp.pca(adata)
```
**Warnings for Code Block - 1**
```
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scmorph/pp/processing.py:215](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scmorph/pp/processing.py#line=214): ImplicitModificationWarning: Modifying `X` on a view results in data being overridden
  adata.X = scaler.transform(np.array(adata.X))
scanpy/preprocessing/_pca/__init__.py:383](http://localhost:8888/scmorph/.venv/lib/python3.13/site-packages/scanpy/preprocessing/_pca/__init__.py#line=382): ImplicitModificationWarning: Setting element `.obsm['X_pca']` of view, initializing view as actual.
  adata.obsm[key_obsm] = X_pca
```
