### 2025-06-30 5dec6e6
KW docstring - clarify removal of features
