### 2025-06-30 5a3b046
remove duplicated sentence
