### 2025-06-30 edbiomedai/scmorph#8 @jeskowagner
> [@jeskowagner](https://github.com/jeskowagner) – I don’t see any mention of the changes described above reflected in the manuscript, and it appears the manuscript has not yet been updated. If I’m overlooking something, please feel free to point me to the exact location where the changes have been made.

My apologies, this specific change got lost in a git conflict and I have re-added the sentences:
Briefly, it removes batch effects using a linear model and thus avoids transformation of features into z-scores. This makes downstream interpretation easier, including allowing for measuring percentage differences rather than differences in z-scores." As always, you can see the current version of the paper [here](https://github.com/edbiomedai/scmorph/blob/paper/paper/paper.pdf).

>It looks like there may be a missing word before “to untrustworthy.” Please let me know if I’m misunderstanding anything.

Sorry about that! I have added the missing word so that it now reads "Note that while it does reduce feature space, its main purpose is to remove untrustworthy features associated with technical confounders.".

Thanks for your time!
