### 2025-06-30 c0e72f1
ignore anndata's implicit modification warnings
