### 2025-06-30 edbiomedai/scmorph#10 @jeskowagner
Hi @tushardave26, thanks for flagging this. These reasons are harmless and can be safely ignored, but it's fair to say that users may be confused about them at first.

The reason why the warnings happen is that, when subsetting an `AnnData` object, it is not copied to conserve memory. This is a good property of `AnnData`. But in some cases, we wish to continue processing a subset of the object, for which it needs to be copied. `ImplicitModificationWarning` warns that the subset `AnnData` object will now be instantiated as an in-memory copy that will no longer be linked to the original, larger object.

In using `scmorph`, we often subset `AnnData` objects, for example to look only at control cells. That leads to the warnings you were observing above. I have chosen to silence the associated warnings, because I believe they cause more confusion here than is useful for our downstream users. (commit was c0e72f14457d4a838ecb390296d7f1de253a7b2e)

All tutorials should now run fine without these warnings.
