### 2025-07-08 edbiomedai/scmorph#7 @jeskowagner
Hi @Armavica, thanks for highlighting that. I meant to use `adata` as a general reference to `AnnData` objects here, but realise that was not clear. I have changed the wording accordingly. The text now [reads](https://scmorph.readthedocs.io/en/latest/tutorials/basics/batch_effects.html):
```
# compute PCA
# the PCA coordinates will be stored in the "X_pca" key of the obsm slot
# meaning you may access them with adata_ctrl_copy.obsm["X_pca"]
sm.pp.pca(adata_ctrl_copy)
```
