### 2025-07-10 782a476
tqdm -> tqdm.auto
