### 2025-07-28 edbiomedai/scmorph#11 @Armavica
Contributing guidelines
Hi, as part of my review of this package that I am doing for JOSS https://github.com/openjournals/joss-reviews/issues/8324 I have to check the following item:
> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Your documentation has a thorough contributing guide that I believe checks point 1., but I couldn't see where points 2. and 3. are addressed. Poking around in the repository, I eventually found https://discourse.scverse.org/ and https://scverse.zulipchat.com/ but these links are difficult to find. Do you think you could feature these forums more prominently in the documentation, and explain in which cases one should go to one or the other in case of issues or to seek support with scmorph?
