+++ docs/references.bib
@@ -40,16 +40,17 @@
   keywords = {methods,normalization,preprocessing,quality control,RNA-seq,scRNA-seq,single-cell}
 }
 
-@article{LinHan2021,
-  title         = {On Boosting the Power of {{Chatterjee}}'s Rank Correlation},
-  author        = {Lin, Zhexiao and Han, Fang},
-  year          = {2021},
-  month         = aug,
-  journal       = {arXiv},
-  eprint        = {2108.06828},
-  doi           = {10/grdrs2},
-  urldate       = {2022-01-21},
-  archiveprefix = {arXiv}
+@article{LinHan2023,
+  title = {On Boosting the Power of {{Chatterjee}}’s Rank Correlation},
+  author = {Lin, Z and Han, F},
+  date = {2023-05-15},
+  journaltitle = {Biometrika},
+  volume = {110},
+  number = {2},
+  pages = {283--299},
+  issn = {0006-3444, 1464-3510},
+  doi = {10.1093/biomet/asac048},
+  langid = {english}
 }
 
 
