### 2025-07-30 bd3109d
add community page to docs
