### 2025-07-31 13e1203
pca: remove whiten, add scale_by_var
