### 2025-07-31 95a99df
add pca scaling to get_ks
