### 2025-08-16 #8324 @csoneson
👋🏻 @jeskowagner I have gone through your submission and overall it looks great - a few suggestions for the paper:

- could you please spell out 'UK' in the affiliations?
- I would suggest to add a sentence explaining what 'morphological profiling' is to the Summary section, to increase accessibility to non-experts
- line 22: averaging or profiles -> averaging of profiles
- line 23: hit-calling -> hit calling (for consistency)
- line 23: recapitulating -> recapitulates or recapitulation of (depending on what it refers to)
- 'anndata' should not be capitalized in the title of Virshup et al (2024) in the bibliography

You can then generate a new proof by typing `@editorialbot generate pdf` in a new comment here.

I will also make another post below with the next steps.
