### 2025-08-17 5207490
Minor paper changes for JOSS
