+++ paper/paper_minimal.bib
@@ -156,7 +156,7 @@
 
 
 @article{VirshupEtAl2024,
-  title   = {Anndata: {{Access}} and Store Annotated Data Matrices},
+  title   = {anndata: {{Access}} and Store Annotated Data Matrices},
   author  = {Virshup, Isaac and Rybakov, Sergei and Theis, Fabian J. and Angerer, Philipp and Wolf, F. Alexander},
   year    = {2024},
   month   = sep,
