### 2025-08-17 #8324 @jeskowagner
The release has been archived at Zenodo: https://doi.org/10.5281/zenodo.16890401
