+++ paper/paper_minimal.bib
@@ -79,7 +79,7 @@
   pages      = {439--452},
   publisher  = {Mary Ann Liebert, Inc., publishers},
   issn       = {1540-658X},
-  doi        = {10/gt2g9c}
+  doi        = {10.1089/adt.2016.726}
 }
 
 @article{RouxdeBezieuxEtAl2024,
