### 2025-08-18 b63ec84
JOSS AEiC paper edits
