+++ paper/paper_minimal.bib
@@ -40,15 +40,19 @@
   archiveprefix = {arXiv}
 }
 
-@article{LinHan2021,
-  title         = {On Boosting the Power of {{Chatterjee}}'s Rank Correlation},
-  author        = {Lin, Zhexiao and Han, Fang},
-  year          = {2021},
-  month         = aug,
-  journal       = {arXiv},
-  eprint        = {2108.06828},
-  doi           = {10.48550/arxiv.2108.06828},
-  archiveprefix = {arXiv}
+@article{LinHan2023,
+  title = {On Boosting the Power of {{Chatterjee}}'s Rank Correlation},
+  author = {Lin, Z and Han, F},
+  year = {2023},
+  month = may,
+  journal = {Biometrika},
+  volume = {110},
+  number = {2},
+  pages = {283--299},
+  issn = {0006-3444, 1464-3510},
+  doi = {10.1093/biomet/asac048},
+  copyright = {https://academic.oup.com/pages/standard-publication-reuse-rights},
+  langid = {english}
 }
 
 
