### 2025-08-21 #8324 @rkurchin
@jeskowagner I think the updated citation is now broken...otherwise things are looking good!
