### 2025-08-23 133fde2
Add JOSS citation
