### 2025-08-23 86615cf
Add paper to Readme
