+++ README.md
@@ -51,3 +51,14 @@ conda install -c conda-forge scmorph
 
 For documentation on the usage of scmorph, please see
 <https://scmorph.readthedocs.io/en/latest/>
+
+
+# Citation
+
+If you use `scmorph` in your work, please cite the `scmorph` publication as follows:
+
+> **scmorph: a Python package for analysing single-cell morphological profiles**
+>
+> Jesko Wagner, Hugh Warden, Ava Khamseh, Sjoerd Viktor Beentjes,
+>
+> _JOSS_ 2025 Aug 23. doi: [10.21105/joss.08324](https://doi.org/10.21105/joss.08324).
