### 2025-08-28 8fe5ee5
Allow internal subsetting of PlateCollection
