### 2025-08-29 1fd09ac
Add plotnine to doc requirements for forthcoming tutorials
