### 2025-08-29 093b9db
Add kwargs doc for qc_images_by_dissimilarity
