### 2025-08-29 7a07d1c
Make image QC notebook discoverable
