+++ docs/tutorials/basics/index.md
@@ -3,6 +3,7 @@
 ```{toctree}
 :maxdepth: 1
 
+image_qc
 batch_effects
 feature_selection
 hit_calling
