### 2025-08-29 0928721
Fix qc notebook links
